backend/requirements.txt
django>=5.0,<5.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
reportlab==4.0.7

